requirements.txt
termcolor
jinja2
jsmin
pyyaml
voluptuous
Pygments
aiohttp
asyncinotify
